Integrates cli.reduce() logic within UmiTemplate.read_idf()
And use UmiTemplate.read_idf in reduce instead
